requirements.txt
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
deprecation==2.1.0
grpcio==1.76.0
idna==3.11
iso8601==2.1.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
prettytable==3.17.0
protobuf==4.25.8
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
schedule==1.2.2
sentry-sdk==2.48.0
six==1.17.0
t-tech-investments==0.3.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
wcwidth==0.2.14
websocket-client==1.9.0
